Fixed double settle bug causing instant game over
Fixed a bug that would sometimes cause settle_blocks() to be called twice (or more times) in a row. This would immediately settle a newly spawned block and block off the spawn area, which results in a game over. The bug is fixed by replacing continue with break in i_check_move_blocks().